+++ src/main/java/com/nexters/momo/member/auth/application/MemberService.java
@@ -5,8 +5,8 @@ import com.nexters.momo.member.auth.domain.MemberRepository;
 import com.nexters.momo.member.auth.exception.DuplicatedUserDeviceIdException;
 import com.nexters.momo.member.auth.exception.DuplicatedUserEmailException;
 import com.nexters.momo.member.auth.presentation.dto.MemberRegisterRequest;
-import com.nexters.momo.member.mypage.common.dto.response.AttendanceDto;
-import com.nexters.momo.member.mypage.common.dto.response.MemberLookUpResponse;
+import com.nexters.momo.member.dto.AttendanceDto;
+import com.nexters.momo.member.dto.MemberLookUpResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
