+++ src/main/java/com/nexters/momo/member/dto/MemberInfoDto.java
@@ -1,4 +1,4 @@
-package com.nexters.momo.member.mypage.common.dto.response;
+package com.nexters.momo.member.dto;
 
 import com.nexters.momo.member.auth.domain.Occupation;
 import com.querydsl.core.annotations.QueryProjection;
